# api/handler/auth.go
// Copyright 2025 Boyuan-IT-Club
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package handler

import (
	"github.com/Boyuan-IT-Club/Meowpick-Backend/application/dto"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/infra/util/token"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/provider"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/types/consts"
	"github.com/gin-gonic/gin"
)

// SignIn godoc
// @Summary 登录
// @Tags auth
// @Accept json
// @Produce json
// @Param body body dto.SignInReq true "SignInReq"
// @Success 200 {object} Response[dto.SignInResp]
// @Router /api/auth/sign_in [post]
func SignIn(c *gin.Context) {
	var err error
	var req dto.SignInReq
	var resp *dto.SignInResp

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}

	tokenStr, _ := token.ExtractToken(c.Request.Header)
	c.Set(consts.CtxToken, tokenStr)

	resp, err = provider.Get().AuthService.SignIn(c, &req)
	PostProcess(c, &req, resp, err)
}

// IsAdmin godoc
// @Summary 是否管理员
// @Description 判断当前用户是否具有管理员权限
// @Tags auth
// @Produce json
// @Success 200 {object} Response[dto.IsAdminResp]
// @Security Bearer
// @Router /api/auth/is_admin [get]
func IsAdmin(c *gin.Context) {
	var err error
	var resp *dto.IsAdminResp

	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().AuthService.IsAdmin(c)
	PostProcess(c, nil, resp, err)
}

// GrantAdmin godoc
// @Summary 授予管理员权限
// @Description 授予指定用户管理员权限
// @Tags auth
// @Accept json
// @Produce json
// @Param body body dto.GrantAdminReq true "GrantAdminReq"
// @Success 200 {object} Response[dto.GrantAdminResp]
// @Security Bearer
// @Router /api/auth/grant_admin [post]
func GrantAdmin(c *gin.Context) {
	var err error
	var req dto.GrantAdminReq
	var resp *dto.GrantAdminResp

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}

	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().AuthService.GrantAdmin(c, &req)
	PostProcess(c, &req, resp, err)
}