# docs/swagger.json
{
    "components": {
        "schemas": {
            "dto.AdminActionVO": {
                "properties": {
                    "action": {
                        "description": "approve/reject/delete",
                        "type": "string"
                    },
                    "actionTime": {
                        "type": "string"
                    },
                    "adminId": {
                        "type": "string"
                    },
                    "adminName": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.ChangeLogVO": {
                "properties": {
                    "action": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "proposalId": {
                        "type": "string"
                    },
                    "targetId": {
                        "type": "string"
                    },
                    "targetType": {
                        "type": "integer"
                    },
                    "updateSource": {
                        "type": "integer"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.CreatorVO": {
                "properties": {
                    "createTime": {
                        "type": "string"
                    },
                    "creatorId": {
                        "type": "string"
                    },
                    "creatorName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.GrantAdminReq": {
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "verifyCode": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.GrantAdminResp": {
                "description": "实际业务数据",
                "properties": {
                    "isAdmin": {
                        "description": "操作后的管理员状态",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "dto.IsAdminResp": {
                "description": "实际业务数据",
                "properties": {
                    "isAdmin": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "dto.ListChangeLogsReq": {
                "properties": {
                    "keyword": {
                        "type": "string"
                    },
                    "page": {
                        "form": "page",
                        "type": "integer"
                    },
                    "pageSize": {
                        "form": "pageSize",
                        "type": "integer"
                    },
                    "type": {
                        "enum": [
                            "course",
                            "proposal",
                            "teacher",
                            "user"
                        ],
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.ListChangeLogsResp": {
                "description": "实际业务数据",
                "properties": {
                    "changeLogs": {
                        "items": {
                            "$ref": "#/components/schemas/dto.ChangeLogVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "dto.ListProposalLogsGroupedResp": {
                "description": "实际业务数据",
                "properties": {
                    "proposals": {
                        "items": {
                            "$ref": "#/components/schemas/dto.ProposalLogVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "dto.ListProposalLogsTimelineResp": {
                "description": "实际业务数据",
                "properties": {
                    "logs": {
                        "items": {
                            "$ref": "#/components/schemas/dto.ProposalTimelineLogVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "dto.ProposalCourseVO": {
                "properties": {
                    "campuses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "category": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "teachers": {
                        "items": {
                            "$ref": "#/components/schemas/dto.TeacherVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.ProposalLogVO": {
                "properties": {
                    "adminAction": {
                        "$ref": "#/components/schemas/dto.AdminActionVO"
                    },
                    "content": {
                        "type": "string"
                    },
                    "course": {
                        "$ref": "#/components/schemas/dto.ProposalCourseVO"
                    },
                    "creator": {
                        "$ref": "#/components/schemas/dto.CreatorVO"
                    },
                    "proposalId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.ProposalSnapshotVO": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "courseName": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.ProposalTimelineLogVO": {
                "properties": {
                    "actionTime": {
                        "type": "string"
                    },
                    "actionType": {
                        "description": "CREATE/APPROVE/REJECT/DELETE/UPDATE/GRANT_ADMIN/REVOKE_ADMIN",
                        "type": "string"
                    },
                    "details": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "logId": {
                        "type": "string"
                    },
                    "operatorId": {
                        "type": "string"
                    },
                    "operatorName": {
                        "type": "string"
                    },
                    "proposalId": {
                        "type": "string"
                    },
                    "proposalSnapshot": {
                        "$ref": "#/components/schemas/dto.ProposalSnapshotVO"
                    }
                },
                "type": "object"
            },
...
            "dto.SignInReq": {
                "properties": {
                    "authId": {
                        "description": "微信开放平台ID",
                        "type": "string"
                    },
                    "authType": {
                        "description": "认证类型(wechat/phone等)",
                        "type": "string"
                    },
                    "verifyCode": {
                        "description": "res.code",
                        "type": "string"
                    }
                },
                "required": [
                    "authId",
                    "authType",
                    "verifyCode"
                ],
                "type": "object"
            },
            "dto.SignInResp": {
                "description": "实际业务数据",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer"
                    },
                    "isAdmin": {
                        "type": "boolean"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.TeacherVO": {
                "properties": {
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_GrantAdminResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GrantAdminResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_IsAdminResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.IsAdminResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_ListChangeLogsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListChangeLogsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_ListProposalLogsGroupedResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListProposalLogsGroupedResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_ListProposalLogsTimelineResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListProposalLogsTimelineResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_SignInResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.SignInResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "选课猫后端接口文档",
        "title": "Meowpick Backend API",
        "version": "1.0"
    },
...
    "paths": {
        "/api/auth/grant_admin": {
            "post": {
                "description": "授予指定用户管理员权限",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.GrantAdminReq",
                                        "summary": "body",
                                        "description": "GrantAdminReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "GrantAdminReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GrantAdminResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "授予管理员权限",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/is_admin": {
            "get": {
                "description": "判断当前用户是否具有管理员权限",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_IsAdminResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "是否管理员",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/auth/sign_in": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.SignInReq",
                                        "summary": "body",
                                        "description": "SignInReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "SignInReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_SignInResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "登录",
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/changelog/list": {
            "post": {
                "description": "按目标类型+ID分页查询变更记录",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.ListChangeLogsReq",
                                        "summary": "req",
                                        "description": "查询参数"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "查询参数",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListChangeLogsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "分页查询变更记录",
                "tags": [
                    "changeLog"
                ]
            }
        },
        "/api/changelog/proposal/grouped": {
            "get": {
                "description": "以提案为维度的分页列表，包含提案基础信息、提议者信息、审核操作信息",
                "parameters": [
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListProposalLogsGroupedResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "按提案聚合的日志列表",
                "tags": [
                    "changeLog"
                ]
            }
        },
        "/api/changelog/proposal/timeline": {
            "get": {
                "description": "一条记录代表一次独立动作的扁平化分页，严格按时间倒序排列",
                "parameters": [
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListProposalLogsTimelineResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "扁平化时间线日志",
                "tags": [
                    "changeLog"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}